+++ src/embed.js
@@ -10,6 +10,8 @@ var scriptSrc = document.currentScript && document.currentScript.src;
 
 var defaultReplayFile = "";
 
+const DEFAULT_REPLAY_BASE = "https://replayweb.page/";
+
 
 // ===========================================================================
 class Embed extends LitElement
@@ -30,6 +32,7 @@ class Embed extends LitElement
     this.coll = "";
     this.paramString = null;
     this.deepLink = false;
+    this.newWindowBase = "";
     this.swInited = false;
     this.embed = null;
     this.reloadCount = 0;
@@ -72,6 +75,9 @@ class Embed extends LitElement
       noSandbox: { type: Boolean },
       noWebWorker: { type: Boolean },
       noCache: { type: Boolean },
+      hideOffscreen: { type: Boolean },
+
+      newWindowBase: { type: String },
 
       errorMessage: { type: String }
     };
@@ -117,6 +123,23 @@ class Embed extends LitElement
     if (this.deepLink) {
       this.updateFromHash();
       window.addEventListener("hashchange", () => this.updateFromHash());
+    }
+
+    this.loadBrowserDefaults();
+  }
+
+  loadBrowserDefaults() {
+    // set defaults for older browsers (eg. Safari < 16) based on recommendations from:
+    // https://lil.law.harvard.edu/blog/2022/09/15/opportunities-and-challenges-of-client-side-playback/
+
+    // likely safari < 16, don't use web workers due to issues with split storage state
+    if (window.GestureEvent !== undefined && window.SharedWorker === undefined ){
+      this.noWebWorker = true;
+    }
+
+    // if no storage manager or estimate, don't cache
+    if (!navigator.storage || !navigator.storage.estimate) {
+      this.noCache = true;
     }
   }
 
@@ -173,8 +196,13 @@ class Embed extends LitElement
         customColl: this.coll,
         config,
         basePageUrl: window.location.href.split("#")[0],
+        baseUrlSourcePrefix: this.newWindowBase,
         embed: this.embed,
       };
+
+      if (!this.deepLink && !params.baseUrlSourcePrefix) {
+        params.baseUrlSourcePrefix = DEFAULT_REPLAY_BASE;
+      }
 
       if (this.noWebWorker) {
         params.noWebWorker = "1";
@@ -182,6 +210,10 @@ class Embed extends LitElement
 
       if (this.noCache) {
         params.noCache = "1";
+      }
+
+      if (this.hideOffscreen) {
+        params.hideOffscreen = "1";
       }
 
       this.paramString = new URLSearchParams(params).toString();